+++ mycodo/config.py
@@ -14,7 +14,7 @@ sys.path.append(os.path.abspath(os.path.dirname(__file__)))
 from config_translations import TRANSLATIONS
 
 MYCODO_VERSION = '8.2.5'
-ALEMBIC_VERSION = '0a8a5eb1be4b'
+ALEMBIC_VERSION = 'a91e6a71028f'
 
 #  FORCE_UPGRADE_MASTER
 #  Set True to enable upgrading to the master branch of the Mycodo repository.